+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/files/Dialogs.kt
@@ -2,12 +2,11 @@ package de.connect2x.messenger.compose.view.files
 
 import androidx.compose.runtime.Composable
 import de.connect2x.trixnity.messenger.util.FileDescriptor
-import de.connect2x.trixnity.messenger.viewmodel.files.PdfDocumentViewModel
 import net.folivo.trixnity.utils.ByteArrayFlow
 
 
 @Composable
-expect fun SaveDialog(
+expect fun SaveFileDialog(
     fileName: String,
     mimeType: String?,
     error: String?,
@@ -16,14 +15,11 @@ expect fun SaveDialog(
 )
 
 @Composable
-expect fun LoadDialog(
+expect fun LoadFileDialog(
     onFileSelect: (FileDescriptor) -> Unit,
     onCloseLoadFileDialog: () -> Unit,
     mode: LoadFileMode,
 )
-
-@Composable
-expect fun PDFReader(documentViewModel: PdfDocumentViewModel, scale: Float = 1f)
 
 enum class LoadFileMode {
     AnyFile,