+++ trixnity-messenger-compose-view/src/webMain/kotlin/de/connect2x/messenger/compose/view/room/settings/ExportRoom.web.kt
@@ -5,6 +5,9 @@ import de.connect2x.trixnity.messenger.export.Destination
 import de.connect2x.trixnity.messenger.export.FileBasedExportRoomProperties
 
 @Composable
-internal actual fun SelectDirectory(properties: FileBasedExportRoomProperties?, result: (Destination?) -> Unit) {
+internal actual fun SelectExportDestination(
+    properties: FileBasedExportRoomProperties?,
+    result: (Destination?) -> Unit
+) {
     // TODO
 }